backend/requirements.txt
asgiref==3.8.1
black==25.1.0
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.0
cryptography==44.0.3
dj-rest-auth==7.0.1
Django==5.2
django-allauth==65.8.0
django-cors-headers==4.7.0
django-filter==25.1
django-rest-knox==5.0.2
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
flake8==7.2.0
idna==3.10
Markdown==3.8
mccabe==0.7.0
mypy_extensions==1.1.0
oauthlib==3.2.2
packaging==25.0
pathspec==0.12.1
pillow==11.2.1
platformdirs==4.3.8
psycopg2-binary==2.9.6
pycodestyle==2.13.0
pycparser==2.22
pyflakes==3.3.2
PyJWT==2.9.0
python-decouple==3.8
requests==2.32.3
requests-oauthlib==2.0.0
sqlparse==0.5.3
tzdata==2025.2
urllib3==2.4.0
